change pidx to unsigned long long and increase PIDX sentinel from unsigned-1 4.29 billion to X40 (0x1<<40)-1 which is more than a trillion

+++ sysrap/sctx.h
@@ -69,7 +69,7 @@ struct sctx
     sevent* evt ;
     const quad2* prd ;
     unsigned idx ;    // local launch index : always zero based
-    unsigned pidx ;   // absolute photon index : offset from zero for launches other than first
+    unsigned long long pidx ;   // absolute photon index : offset from zero for launches other than first
 
 #if !defined(PRODUCTION) && defined(DEBUG_PIDX)
     bool pidx_debug ;
